### 2021-11-03 #3886 @whedon
[PRE REVIEW]: NeuralFieldEq.jl: A flexible solver to compute Neural Field Equations in several scenarios
**Submitting author:** @tiagoseq (<a href="http://orcid.org/0000-0001-8579-3676">Tiago Sequeira</a>)
**Repository:** <a href="https://github.com/tiagoseq/NeuralFieldEq.jl" target ="_blank">https://github.com/tiagoseq/NeuralFieldEq.jl</a>
**Version:** v0.1.2
