### 2021-11-03 #3886 @danielskatz
@tiagoseq - thanks for your submission.  Due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS.  You should hear back in a week or so.

In addition, you could work on the possibly missing DOI that whedon suggests and the incorrect DOI, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
