### 2021-11-03 #3886 @tiagoseq
Hello,
Some time ago, Nicolas Rougier suggested that I should submit the solver to
JOSS and indicate him as a reviewer. Because the method implemented is
based on the novel technique presented in this paper
<https://hal.inria.fr/hal-00872132> authored by Axel Hutt and Nicolas
Rougier.

This solver handles with several types of Neural Field equations, being
user friendly, easy to use and efficient as possible. As far as I can tell,
there isn't other piece of software that brings together this features.

I will check the issues with DOI and update the .bib file.

Thank you,
Tiago

Em qua., 3 de nov. de 2021 às 13:35, Daniel S. Katz <
***@***.***> escreveu:

> @tiagoseq <https://github.com/tiagoseq> - thanks for your submission. Due
> to the relatively small amount of code, the editors will now discuss if it
> meets the substantial scholarly effort
> <https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort>
> criterion for review by JOSS. You should hear back in a week or so.
>
> In addition, you could work on the possibly missing DOI that whedon
> suggests and the incorrect DOI, but note that some may be incorrect. Please
> feel free to make changes to your .bib file, then use the command @whedon
> check references to check again, and the command @whedon generate pdf
> when the references are right to make a new PDF. Whedon commands need to be
> the first entry in a new comment.
>
