+++ JOSS/paper.bib
@@ -5,7 +5,7 @@ journal = {The Journal of Physiology},
 volume = {117},
 number = {4},
 pages = {500-544},
-doi = {https://doi.org/10.1113/jphysiol.1952.sp004764},
+doi = {10.1113/jphysiol.1952.sp004764},
 year = {1952}
 }
 
@@ -92,13 +92,14 @@ doi = {10.1186/2190-8567-4-1}
 }
 
 @article{Amari,
-   author  = {S. L. Amari},
-   title   = {Dynamics of pattern formation in lateral.inhibition type neural fields},
-   journal = {Biol. Cybernet},
-   year    = {1977},
-   volume  = {27},
-   pages   = {77-87},
-   note    = {}
+author = {Amari, Shun-ichi},
+year = {1977},
+month = {09},
+pages = {77-87},
+title = {Dynamic of pattern formation in lateral-inhibition type neural fields},
+volume = {27},
+journal = {Biological cybernetics},
+doi = {10.1007/BF00337259}
 }
 
 @article{SequeiraLima,
