### 2021-12-01 #3886 @jarvist
OK! Let's ask Nicolas directly here, as he seems to be a regular user of GitHub.

We need at least two reviewers for JOSS. As Nicolas appears to be a domain expert, I suppose someone with PDE / Julia experience would be complementary.
