### 2021-12-01 #3886 @rougier
I can review but I might be a bit late (note before end of December I would say)
