### 2021-12-03 #3974 @whedon
[REVIEW]: NeuralFieldEq.jl: A flexible solver to compute Neural Field Equations in several scenarios
**Submitting author:** @tiagoseq (<a href="http://orcid.org/0000-0001-8579-3676">Tiago Sequeira</a>)
**Repository:** https://github.com/tiagoseq/NeuralFieldEq.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
