### 2022-01-03 tiagoseq/NeuralFieldEq.jl#29 @rougier
Plotting issue
I'm doing the [review](https://github.com/openjournals/joss-reviews/issues/3974) for JOSS and when I try the example on the README (Neural field with breather type instabilities), I got:

```
...
julia> # Plotting
       # Animate the delayed solutions
       anim = @animate for i = 1:length(tj)
         p1 = plot(x,y,V_v5(i),st=:surface,title="v=5. t=$(tj[i])",color=:balance,zlims=(-2.8,0.95))
         p2 = plot(x,y,V_v3(i),st=:surface,title="v=3. t=$(tj[i])",color=:balance,zlims=(-2.8,0.95))
         plot(p1,p2)
       end
ERROR: UndefVarError: x not defined
Stacktrace:
 [1] macro expansion
   @ ./REPL[20]:4 [inlined]
 [2] top-level scope
   @ ~/.julia/packages/Plots/FI0vT/src/animation.jl:197

julia> gif(anim, "v5v3_breather.gif",fps=10)
ERROR: UndefVarError: anim not defined
Stacktrace:
 [1] top-level scope
   @ REPL[21]:1
```

I'm not too familiar with julia and I just copy paste the code to julia interpreter.
