### 2022-01-03 #3974 @rougier
Here are some temporary comments on the submission since I'm waiting for an [issue](https://github.com/tiagoseq/NeuralFieldEq.jl/issues/29) to be solved before going further.

1. Repository is way too big (36Mo) due mostly to the presence of gif animations in the imag and doc directories. Problem is that the doc directories containes the build website for no obvious reason. I would advise to cleanup the repository. But problem will remain because the build addition has been commited and the repository will stay big unless the git history is rewritten. Another solution would be to create a new repository and make sure to only add necessary stuff.

2. The name of the repository is a bit weird with the .jl extension but I don't know if it is usage for Julia packages.

3. Since computation is a bit long, would it be possible to have some kind of progress indicator when it is ran?

4. The summary in the paper needs to be worked a bit. Especially the first sentence that seems to be missing a verb. But more generally, I think i would be ok to dive directly in the neural field, starting at "Wilson & Cowan (1972)...".

5. Equation 1 seems to be missing the axonal speed term. This is ok since it does not appear in the original equation but you shoudl rephrase the paragraph to explain that speed can be introduced in the equation and then rewrite the full equation.

5. "when the coupling is positive (negative) the synapses are excitatory (inhibitory)" -> "when the coupling is positive (resp. negative), the synapses are excitatory (resp. inhibitory)

6. "Although the this" -> "Although this"

7. Overall, I think the paper could be improved because there are some typos that can easily be fixed and the package usage section should be improved to better introduce the package. As it is currently written it's a bit hard to follow author's explanations.
