### 2022-01-03 ca92d20
Update README.md
Bug fix in example Neural field with breather type instabilities
