### 2022-01-03 tiagoseq/NeuralFieldEq.jl#29 @tiagoseq
Hello,
I'm sorry, two instructions were missing before constructing the animation, namely "x = V_v5.x" and "y = V_v5.y". I've already updated the readme file.

I kindly ask you to run the script again to see if the animations are generated (it will take a few minutes do generate the .gif file). If so, i'll close the issue.

Thank you,
Tiago
