### 2022-01-03 #3974 @tiagoseq
Hello,

I can answer some of the comments:

1. Indeed the repo is too big due to the animations present in the docs section. I'll see what I can do to solve this problem, I guess I will take your advice and create a new repository for the docs, or re-write "lighter" documentation without too many animations.
I have to address this question carefully, I'll give updates on this soon.

2. The name of the repository follows the julia guidelines about package names, that's why the extension .jl appears. For example, these packages have the extension: [DifferentialEquations.jl](https://github.com/SciML/DifferentialEquations.jl) and [FFTW.jl](https://github.com/JuliaMath/FFTW.jl).

3. I guess that is possible to add a progress indicator, I'll give updates soon.

4. - 7. Since these issues are easier to fix, I'll check them first. I agree with you, the summary needs to be written more carefully.

8. Ok, I will read carefully again the package usage section and try to improve it to make it more clear.

When I have some of these points done, I'll write here.

Thanks,
Tiago
