### 2022-01-03 #3974 @gdetor
Here are my comments on the manuscript. I'll provide my comments on the source code later.

Text Comments
-----------------

1. Memory, perception, and other functions of the human brain do not occur only in the cerebral cortex. Please rephrase the sentences in lines 7 - 10.

2. It would be better to rewrite equation 1 in a more general way. For instance, the kernel function K(x, t) is not always isotropic or homogeneous. You could incorporate the delay term as well. Why is the norm a Euclidean one (is the only supported norm)? The integral in equation 1 is always two-dimensional (please replace the square with d in the differential of the integral). You could also use a different symbol for indicating the dimension. What about 3-dimensional neural fields?

3. You could consider citing these two references: Coombes, "Waves, bumps, and patterns in neural field theories", 2005 and Bressloff, "Spatiotemporal dynamics of continuum neural fields", 2011. Thus, readers who are not experts in neural fields can get a general idea.

4. Could you please provide a reference regarding the performance of Julia? See lines 40-42 in the main text.

5. The first paragraph in the "Package usage" section is obscure. This section needs some improvements.

6. The sentence in lines 44-46 is not clear.

7. Could you provide more details regarding the performance of the code?

Language
----------
The text needs polishing. Here are some typos I spotted in the main text:
1. Please remove "that" from the very first sentence of the summary.
2. Line 30, "Although the this ..." -> "Although this ..."
3. Lines 35-36, "the scope of the master's thesis [Ref] ..." -> "the scope of [Ref]"
4. Line 42, "the the" -> "the"
5. Line 44, "non-delayed or delayed equations" -> "non-delayed or delayed"
6. Line 58, "access it at the previously" -> "access it at previously"
