### 2022-01-04 tiagoseq/NeuralFieldEq.jl#29 @rougier
Thanks, it's working now. By the way for the animation, you could set the value of vmin/vmax such as to have a fixed colorbar that would ease the reading of the animation. Else, colors are constantly changing making it difficult to see what is happening.
