### 2022-01-04 tiagoseq/NeuralFieldEq.jl#29 @tiagoseq
In the documentation the same example is addressed and when plotting the graphs you can notice the arguments `zlims=(-0.43,0.54)` and `ylims=(-0.43,0.54)`, for the surface plot and the line plot respectively. If you use these arguments to fix the limits of the plots the color bar will maintain during the simulation.
Unfortunately in the example that i mentioned, the colorbar is hidden, but if the user plot only a surface (without any subplot) the colorbar appears by default.
