+++ JOSS/paper.md
@@ -20,11 +20,11 @@ bibliography: paper.bib
 
 # Summary
 
-Since @hodgkinhuxley that mathematical modelling proved to be a prominent neuroscience research area. One of its main challenges is to describe phenomena like memory, perception, etc. These neurobiological processes occur in the cortex, a layer of the brain that contains approximately $10^{10}$ neurons distributed along $2500\,cm^2$ of surface area with $2.8\,mm$ of depth. Driven by such complex structure, @WilsonCowan and @Amari derived a model called Neural Field Equation (NFE), that treats the cerebral cortex as a continuum space being able to deal with these large scale dynamical neuronal patterns. The equation has the following form
+Since @hodgkinhuxley mathematical modelling proved to be a prominent neuroscience research area. One of its main challenges is to describe phenomena like memory, perception, etc. These neurobiological processes occur in the cortex, a layer of the brain that contains approximately $10^{10}$ neurons distributed along $2500\,cm^2$ of surface area with $2.8\,mm$ of depth. Driven by such complex structure, @WilsonCowan and @Amari derived a model called Neural Field Equation (NFE), that treats the cerebral cortex as a continuum space being able to deal with these large scale dynamical neuronal patterns. The equation has the following form
 \begin{equation}\label{eq:NFE}
   \alpha \frac{\partial V}{\partial t}\left(\mathbf{x},t\right) = I\left(\mathbf{x},t\right) - V\left(\mathbf{x},t\right) + \int_{\Omega} K\left(||\mathbf{x}-\mathbf{y}||_2\right)S\big[V\left(\mathbf{y},t\right)\big]\,\,d^2\mathbf{y},
 \end{equation}
-where $\Omega=[-\frac{L}{2},\frac{L}{2}]^d$ with $d=1,2$; $V(\mathbf{x},t)$ is the membrane potential at point $\mathbf{x} \in \Omega$ at time $t$; $I(\mathbf{x},t)$ is the external input applied to the neural field; $K\left(||\mathbf{x}-\mathbf{y}||_2\right)$ is the average strength of connectivity between neurons located at points $\mathbf{x}$ and $\mathbf{y}$, when the coupling is positive (negative) the synapses are excitatory (inhibitory); $S(V)$ is the firing rate function, which converts the potential to the respective firing rate result; And $\alpha$ is the constant decay rate.
+where $\Omega=[-\frac{L}{2},\frac{L}{2}]^d$ with $d=1,2$; $V(\mathbf{x},t)$ is the membrane potential at point $\mathbf{x} \in \Omega$ at time $t$; $I(\mathbf{x},t)$ is the external input applied to the neural field; $K\left(||\mathbf{x}-\mathbf{y}||_2\right)$ is the average strength of connectivity between neurons located at points $\mathbf{x}$ and $\mathbf{y}$, when the coupling is positive (resp. negative) the synapses are excitatory (resp. inhibitory); $S(V)$ is the firing rate function, which converts the potential to the respective firing rate result; And $\alpha$ is the constant decay rate.
 
 Since then there have been endeavours to improve NFEs. For example, depending on the cortical region the axonal speed can vary between $100\,m/s$ to $1\,m/s$, so considering the time spent by the stimulus to travel from neurons at $y$ to the ones at $x$ leads us to formulate more accurate models. Or the noisy neuronal interactions, generated by the stochastic nature of the neuron or from extrinsic noise sources that can arise from inputs of other neuronal networks.
 
@@ -32,11 +32,11 @@ Since then there have been endeavours to improve NFEs. For example, depending on
 
 # Statement of need
 
[…]
