+++ JOSS/paper.md
@@ -20,38 +20,66 @@ bibliography: paper.bib
 
 # Summary
 
-Since @hodgkinhuxley mathematical modelling proved to be a prominent neuroscience research area. One of its main challenges is to describe phenomena like memory, perception, etc. These neurobiological processes occur in the cortex, a layer of the brain that contains approximately $10^{10}$ neurons distributed along $2500\,cm^2$ of surface area with $2.8\,mm$ of depth. Driven by such complex structure, @WilsonCowan and @Amari derived a model called Neural Field Equation (NFE), that treats the cerebral cortex as a continuum space being able to deal with these large scale dynamical neuronal patterns. The equation has the following form
+In their works, @WilsonCowan presented a model, later developed by @Amari, named Neural Field Equations (NFE). @Bressloff:minibook and @Coombes:WavesBumps, summarise the main theoretical results and enumerate a wide variety of neurological phenomena that can be described by these equations. Such as working memory [@Laing:WM], travelling waves, etc. The model in study is given by
 \begin{equation}\label{eq:NFE}
-  \alpha \frac{\partial V}{\partial t}\left(\mathbf{x},t\right) = I\left(\mathbf{x},t\right) - V\left(\mathbf{x},t\right) + \int_{\Omega} K\left(||\mathbf{x}-\mathbf{y}||_2\right)S\big[V\left(\mathbf{y},t\right)\big]\,\,d^2\mathbf{y},
+  \alpha \frac{\partial V}{\partial t}\left(\mathbf{x},t\right) = I\left(\mathbf{x},t\right) - V\left(\mathbf{x},t\right) + \int_{\Omega} K\left(||\mathbf{x}-\mathbf{y}||_2\right)S\big[V\left(\mathbf{y},t\right)\big]\,\,d^k\mathbf{y},
 \end{equation}
-where $\Omega=[-\frac{L}{2},\frac{L}{2}]^d$ with $d=1,2$; $V(\mathbf{x},t)$ is the membrane potential at point $\mathbf{x} \in \Omega$ at time $t$; $I(\mathbf{x},t)$ is the external input applied to the neural field; $K\left(||\mathbf{x}-\mathbf{y}||_2\right)$ is the average strength of connectivity between neurons located at points $\mathbf{x}$ and $\mathbf{y}$, when the coupling is positive (resp. negative) the synapses are excitatory (resp. inhibitory); $S(V)$ is the firing rate function, which converts the potential to the respective firing rate result; And $\alpha$ is the constant decay rate.
+with $\Omega=[-\frac{L}{2},\frac{L}{2}]^k$ with $k=1,2$ and $||.||_2$ the euclidean norm; $V(\mathbf{x},t)$ is the membrane potential at point $\mathbf{x} \in \Omega$ at time $t$; $I(\mathbf{x},t)$ is the external input applied to the neural field; $K\left(||\mathbf{x}-\mathbf{y}||_2\right)$ is the average strength of connectivity between neurons located at points $\mathbf{x}$ and $\mathbf{y}$ in a homogeneous neural field. When the coupling is positive (resp. negative) the synapses are excitatory (resp. inhibitory); $S(V)$ is the firing rate function, which converts the potential to the respective firing rate result; And $\alpha$ is the constant decay rate. The equation is coupled with the initial condition $V(\mathbf{x},0) = V_0(\mathbf{x})$.
 
[…]
