+++ JOSS/paper.bib
@@ -1,14 +1,3 @@
-@article{hodgkinhuxley,
-author = {Hodgkin, A. L. and Huxley, A. F.},
-title = {A quantitative description of membrane current and its application to conduction and excitation in nerve},
-journal = {The Journal of Physiology},
-volume = {117},
-number = {4},
-pages = {500-544},
-doi = {10.1113/jphysiol.1952.sp004764},
-year = {1952}
-}
-
 @article{fftw,
 author={M. {Frigo} and S. G. {Johnson}},
 journal={Proceedings of the IEEE},
@@ -19,24 +8,6 @@ number={2},
 pages={216-231},
 doi={10.1109/JPROC.2004.840301}}
 
-@misc{cortex,
-  author= {Eric Chudler},
-  title = {Brain Facts and Figures},
-  howpublished = {\url{http://faculty.washington.edu/chudler/facts.html}},
-  note = {Accessed: 2021-02-22}
-}
-
-@article{nneuronios,
-author = {Azevedo, Frederico and Carvalho, Ludmila and Grinberg, Lea and Farfel, Jose and Ferretti-Rebustini, Renata and Leite, Renata and Filho, Wilson and Lent, Roberto and Herculano-Houzel, Suzana},
-year = {2009},
-month = {04},
-pages = {532-41},
-title = {Equal Numbers of Neuronal and Nonneuronal Cells Make the Human Brain an Isometrically Scaled-Up Primate Brain},
-volume = {513},
-journal = {The Journal of comparative neurology},
-doi = {10.1002/cne.21974}
-}
-
 @article{WilsonCowan,
 author = {Wilson, Hugh and Cowan, Jack},
 year = {1972},
@@ -46,17 +17,6 @@ title = {Excitatory and Inhibitory Interactions in Localized Populations of Mode
 volume = {12},
 journal = {Biophysical journal},
 doi = {10.1016/S0006-3495(72)86068-5}
-}
-
-@article{HuttRougier:1,
-author = {Hutt, Axel and P. Rougier, Nicolas},
-year = {2010},
-month = {11},
-pages = {055701},
-title = {Activity spread and breathers induced by finite transmission speeds in two-dimensional neural fields},
-volume = {82},
-journal = {Physical review. E, Statistical, nonlinear, and soft matter physics},
-doi = {10.1103/PhysRevE.82.055701}
 }
 
 @article{HuttRougier:2,
@@ -78,17 +38,6 @@ title = {Neural Field Simulator: two-dimensional spatio-temporal dynamics involv
 volume = {9},
 journal = {Frontiers in Neuroinformatics},
 doi = {10.3389/fninf.2015.00025}
-}
-
-@article{RiedlerChristian,
-author = {Kuehn, Christian and G Riedler, Martin},
-year = {2014},
-month = {04},
-pages = {1},
-title = {Large Deviations for Nonlocal Stochastic Neural Fields},
-volume = {4},
-journal = {Journal of mathematical neuroscience},
-doi = {10.1186/2190-8567-4-1}
 }
 
 @article{Amari,
@@ -132,3 +81,69 @@ title = {Numerical simulations of one- and two-dimensional stochastic neural fie
 school = {Instituto Superior Técnico - Universidade de Lisboa},
 year = {2021},
 note = {\url{https://fenix.tecnico.ulisboa.pt/cursos/mma/dissertacao/1691203502344856}}}
+
+@article{Coombes:WavesBumps,
+author = {Coombes, Stephen},
+year = {2005},
+month = {09},
+pages = {91-108},
+title = {Waves, bumps, and patterns in neural field theories. Biol. Cybern. 93(2), 91-108},
+volume = {93},
[…]
