+++ JOSS/paper.md
@@ -20,17 +20,11 @@ bibliography: paper.bib
 
 # Summary
 
-In their works, @WilsonCowan presented a model, later developed by @Amari, named Neural Field Equations (NFE). @Bressloff:minibook and @Coombes:WavesBumps, summarise the main theoretical results and enumerate a wide variety of neurological phenomena that can be described by these equations. Such as working memory [@Laing:WM], travelling waves, etc. The model in study is given by
-\begin{equation}\label{eq:NFE}
-  \alpha \frac{\partial V}{\partial t}\left(\mathbf{x},t\right) = I\left(\mathbf{x},t\right) - V\left(\mathbf{x},t\right) + \int_{\Omega} K\left(||\mathbf{x}-\mathbf{y}||_2\right)S\big[V\left(\mathbf{y},t\right)\big]\,\,d^k\mathbf{y},
-\end{equation}
-with $\Omega=[-\frac{L}{2},\frac{L}{2}]^k$ with $k=1,2$; $V(\mathbf{x},t)$ is the membrane potential at point $\mathbf{x} \in \Omega$ at time $t$; $I(\mathbf{x},t)$ is the external input applied to the neural field; $K\left(||\mathbf{x}-\mathbf{y}||_2\right)$ is the average strength of connectivity between neurons located at points $\mathbf{x}$ and $\mathbf{y}$ in a homogeneous neural field. When the coupling is positive (resp. negative) the synapses are excitatory (resp. inhibitory); $S(V)$ is the firing rate function, which converts the potential to the respective firing rate result; And $\alpha$ is the constant decay rate. The equation is coupled with the initial condition $V(\mathbf{x},0) = V_0(\mathbf{x})$.
-
-One of the improvements to consider in \autoref{eq:NFE}, is to take into account the time taken by the stimulus to travel from neurons at $\mathbf{y}$ to the ones at $\mathbf{x}$, leading to the delayed equation,
+In their works, @WilsonCowan presented a model, later developed by @Amari, named Neural Field Equations (NFE). @Bressloff:minibook and @Coombes:WavesBumps, summarise the main theoretical results and enumerate a wide variety of neurological phenomena that can be described by these equations. Such as working memory [@Laing:WM], travelling waves, etc. One of the improvements to consider in the simpler NFE, is to take into account the time taken by the stimulus to travel from neurons at different points. This leads to the delayed equation,
 \begin{equation}\label{eq:dNFE}
   \alpha \frac{\partial V}{\partial t}\left(\mathbf{x},t\right) = I\left(\mathbf{x},t\right) - V\left(\mathbf{x},t\right) + \int_{\Omega} K\left(||\mathbf{x}-\mathbf{y}||_2\right)S\big[V\left(\mathbf{y},t-d\left(\mathbf{x},\mathbf{y}\right)\right)\big]\,\,d^k\mathbf{y},
 \end{equation}
-where $d\left(\mathbf{x},\mathbf{y}\right)=\frac{||\mathbf{x}-\mathbf{y}||_2}{v}$ represents the delay, which is assumed to only depend on the distance and on the transmission speed $v$. If $v$ is sufficiently high, $d$ can be neglected and \autoref{eq:dNFE} is reduced to \autoref{eq:NFE}. Moreover, $V$ satisfies the initial condition $V(\mathbf{x},t_0)=V_0(\mathbf{x},t_0),\,\,t_0\in\left[-\tau_{max},\,\,0\right]$, wherein $\tau_{max}$ is the maximum delay correspondent to $\Omega$.
[…]
