### 2022-01-26 #3974 @tiagoseq
Hello,
I've released a new major version of the manuscript, with the following changes:
- Reformulated the Package usage section, i think now is more straight to the point. (both reviewers suggestion @rougier and @gdetor);
- The code example had some modifications;
- Introduced the delay term in the equation (both reviewers suggestion), as well as the delayed stochastic NFE;
- Added more context and references to the Neural Field Equations (@gdetor suggestion) in Summary section;
- Introduced new paragraph with two new references in Statement of need section, to give more context about the importance of rapid and efficient solvers;
- Opened a new section called Example and code performance, and moved the example to this section and added a table with benchmarks for the 1D and 2D problems for several spatial nodes;
- Polished the text and corrected the typos.

Now, answering some questions:
@rougier
1. Regarding the repo size, i've checked the settings of the repo on github and says that is about 10 Mb. Am I seeing this wrong? Yet, I have to work in this issue;
2.  I thought that add a status bar would be straightforward, but it isn't. Probably, next week I have to check it out.

@gdetor
 1. I wrote an isotropic kernel because the solver do not support (yet) such functions. In this method K is homogeneous and isotropic (check line 15);
 2. Yes, the euclidean norm is the only supported, the method implemented only considers p-2 norms;
 3. I did not addressed the 3D case, because the first motivation of the solver was to study fields that exhibit working memory mechanisms applied to cognitive robotics (first paragraph of Statement of need) and there was no need to have one more dimension.

Thank you, and I'll be grateful when you can give feedback,
Tiago
