### 2022-01-31 #3974 @rougier
@tiagoseq

You can check the size locally with command `du -sh NeuralFieldEq.jl` (Linux or OSX). I think size reported by GitHub is only for current files, without the history of commits which can be quite large.

For the progress bar, it's only a suggestion, nothing really important. There is certainly a Julia package to do just that. If it's the case that would be nice and easy to add it.
