### 2022-01-31 #3974 @gdetor
@tiagoseq @jarvist
Here are my comments on the Source Code:
1. Many parts of the code are not documented (for instance the functions in script AuxScripts.jl or in SolveStruct.jl). Please document all the functions and variables within the code.
2. Please provide a list with the contributors (on the Github repository as a section in the README file)
3. Please add community guidelines (especially for issues and bugs) in the repo's README file.

Other than that everything looks fine to me. The installation instructions are OK. The code works properly. The performance of the package is now reported in the main text. The repository contains automated tests and all tests pass with no any apparent issues.

My previous comments regarding the main text have been addressed by the author.

@tiagoseq Could you please generate the new pdf article proof?
