### 2022-02-07 #3974 @tiagoseq
@gdetor
- During this week I'll provide the documentation for the auxiliary functions and the functors present in the SolveStruct.jl and others that might need more insights and explanation.
- Regarding the contributors to the package, (until now) I was the one responsible for writing it. But I'll specify that in the README file as you suggested.
- Yes, I'll write the community guidelines.

The new version of the manuscript is generated.

@rougier
I will move the docs to an independent repository and clean the git history to reduce the size of the repo. I'll look this issue after wednesday. Thank you for this suggestion!
Yes, there is a Julia package to implement a status bar, probably is straight forward. I like that nice-to-have feature, I'll try to implement it after addressing the major issues pointed out!

Thank you!
