### 2022-02-15 cd77e30
Documented auxiliary functions and functors
