### 2022-02-16 da1f199
undo functors documentation
