+++ src/SolveStructMethods1D.jl
@@ -3,11 +3,6 @@
 
 # Return V(x,ti)
 function (a::SolveOutDet1D)(ti::AbstractFloat)
-    @doc raw"""
-        a(ti)
-
-    This functor returns the 1D deterministic solution at instant `ti`
-    """
     t = findlast(i->i==ti,a.tsaved)
     return a.V[(t-1)*a.N+1:a.N*t]
 end
@@ -15,11 +10,6 @@ end
 
 # Return V(xj,ti)
 function (a::SolveOutDet1D)(xj::AbstractFloat,ti::AbstractFloat)
-    @doc raw"""
-        a(xj,ti)
-
-    This functor returns the 1D deterministic solution at point `xj` and instant `ti`
-    """
     t  = findlast(i->i==ti,a.tsaved)
     x  = findlast(j->j==xj,a.x)
     Vi = @view a.V[(t-1)*a.N+1:a.N*t]
@@ -36,11 +26,6 @@ end
 
 # Return V(x,ti,p)
 function (a::SolveOutSto1D)(ti::AbstractFloat,p::Int)
-    @doc raw"""
-        a(ti,p)
-
-    This functor returns the path `p` at instant `ti` (1D domain)
-    """
     t = findlast(i->i==ti,a.tsaved)
     return a.V[(t-1)*a.N+1:a.N*t,p]
 end
@@ -48,11 +33,6 @@ end
 
 # Return Vmean(x,ti)
 function (b::SolveOutSto1D)(ti::AbstractFloat)
-    @doc raw"""
-        b(ti)
-
-    This functor returns the 1D mean stochastic solution at instant `ti`
-    """
     t = findlast(i->i==ti,b.tsaved)
     return b.meanV[(t-1)*b.N+1:b.N*t]
 end
@@ -60,11 +40,6 @@ end
 
 # Return V(xj,ti,p)
 function (a::SolveOutSto1D)(xj::AbstractFloat,ti::AbstractFloat,p::Int)
-    @doc raw"""
-        a(xj,ti,p)
-
-    This functor returns the path `p` at point `xj` and instant `ti` (1D domain)
-    """
     t  = findlast(i->i==ti,a.tsaved)
     x  = findlast(j->j==xj,a.x)
     Vi = @view a.V[(t-1)*a.N+1:a.N*t,p] # path at ti
@@ -78,11 +53,6 @@ end
 
 # Return Vmean(xj,ti)
 function (b::SolveOutSto1D)(xj::AbstractFloat,ti::AbstractFloat)
-    @doc raw"""
-        b(xj,ti)
-
-    This functor returns the 1D mean stochastic solution at point `xj` and instant `ti`
-    """
     t  = findlast(i->i==ti,b.tsaved)
     x  = findlast(j->j==xj,b.x)
     Vi = @view b.meanV[(t-1)*b.N+1:b.N*t]
