### 2022-02-16 3da11d8
Corrected documentation in aux functions
