+++ src/solveNFE1D.jl
@@ -69,6 +69,7 @@ function solveNFE(prob::ProbOutput1D,saveat) # Deterministic method for 1D
     a   = similar(svu) # rfft of the integral operator A(X,t)
     j   = 1 # Index for saveat
 
+    prog = Progress(length(t), desc="Computing: ", color=:blue) # Define progress bar
     @inbounds for i = 1:length(t) # Time loop
         ti = t[i]
         # Store solution V at saveat instants
@@ -95,7 +96,8 @@ function solveNFE(prob::ProbOutput1D,saveat) # Deterministic method for 1D
         # Every delay block moves one block down. The last one is deleted
         sv[hN+1:end] .= @view sv[1:hN*(rings-1)] 
         mul!(svu,P,prob.S.(V)) # Apply the Real Fourier Transform to V
-        sv[1:hN] .= svu        # Update sv first delay ring
+        sv[1:hN] .= svu        # Upd sv first delay ring
+        next!(prog) #upd status bar
     end #end time loop
 
     # Build our solution structure output
@@ -136,6 +138,7 @@ function solveNFE(prob::ProbOutput1D,saveat,ϵ,np,ξ=0.1)
     # Shift λ due to the also shifted output of fft
     c = N*sqrt(dt)*ϵ*fftshift(λ) # ϵ is the level of additive noise
     
+    prog = Progress(np, desc="Paths computed: ", color=:blue) # Define progress bar
     # Trajectories loop
     @inbounds for p = 1:np
         copy!(V,prob.V0)  # Initial condition V(X,0) = V0 (Natural domain)
@@ -173,7 +176,7 @@ function solveNFE(prob::ProbOutput1D,saveat,ϵ,np,ξ=0.1)
             mul!(svu,P,prob.S.(V)) # Apply the Real Fourier Transform to V
             sv[1:hN,:].= svu       # Update sv first delay ring
         end #end time loop
-
+        next!(prog) # upd status bar next trajectory
     end #end trajectories loop
 
     # Compute the mean solution across all trajectories
