### 2022-02-18 d8eb3be
CompatHelper: add new compat entry for ProgressMeter at version 1, (keep existing compat)
