+++ Project.toml
@@ -12,6 +12,7 @@ ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 [compat]
 Distributions = "^0.25"
 FFTW = "^1.3"
+ProgressMeter = "1"
 julia = "1.5"
 
 [extras]
