### 2022-02-18 #3974 @tiagoseq
Hello,
Sorry for the delay, today I finally had time to work on this issue.
@rougier yes, I rebuilt the PDF with the comments that you and @gdetor made to the manuscript. Every comment, suggestion was addressed in this new version.

As for the code, what is already done is:
- I added the status when executing the function `solveNFE` @rougier
- Updated readme with a new section that contains the contributors, and community guidelines @gdetor
- Documented structures `ProbStruct.jl` and `SolveStruct.jl`, the module `AuxFunctions.jl` is documented and `DomainConstruct.jl`.

What I'll do next @rougier :
- Move documentation to other independent repository
- Clean git commit history
- Release a new version of the software.

When the new version is released, I'll communicate here
