+++ Project.toml
@@ -1,0 +1,21 @@
+name = "NeuralFieldEq"
+uuid = "d4f2cf9b-1959-46fe-97e6-985c98850872"
+authors = ["Tiago Sequeira"]
+version = "0.1.2"
+
+[deps]
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
+
+[compat]
+Distributions = "^0.25"
+FFTW = "^1.3"
+julia = "1.5"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]

[…]
