+++ Project.toml
@@ -1,7 +1,7 @@
 name = "NeuralFieldEq"
 uuid = "d4f2cf9b-1959-46fe-97e6-985c98850872"
 authors = ["Tiago Sequeira"]
-version = "0.1.2"
+version = "0.1.3"
 
 [deps]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
