+++ Project.toml
@@ -11,6 +11,7 @@ ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 
 [compat]
 Distributions = "^0.25"
+ProgressMeter = "^1.7.1"
 FFTW = "^1.3"
 julia = "1.5"
 
