### 2022-02-21 #3974 @tiagoseq
Hello,
A pull request with a new version with the following features has been made:
- Status bar when executing function `solveNFE`
- New independent repository, [NFEdocs.jl](https://tiagoseq.github.io/NFEdocs.jl/dev/) to host the documentation of the package
- Documented structures and the auxiliary functions module
- Cleaned git commit history by creating an orphan branch with the latest commit and files, deleting the master branch and renaming the orphan branch to master. (However, I have to check this after the auto-merge)

I'm waiting for the [auto-merge](https://github.com/JuliaRegistries/General/pull/55127), made by JuliaRegistrator autobot. When the new version (v0.1.7) is released I'll let you know.
