+++ JOSS/paper.bib
@@ -1,0 +1,149 @@
+@article{fftw,
+author={M. {Frigo} and S. G. {Johnson}},
+journal={Proceedings of the IEEE},
+title={The Design and Implementation of FFTW3},
+year={2005},
+volume={93},
+number={2},
+pages={216-231},
+doi={10.1109/JPROC.2004.840301}}
+
+@article{WilsonCowan,
+author = {Wilson, Hugh and Cowan, Jack},
+year = {1972},
+month = {02},
+pages = {1-24},
+title = {Excitatory and Inhibitory Interactions in Localized Populations of Model Neurons},
+volume = {12},
+journal = {Biophysical journal},
+doi = {10.1016/S0006-3495(72)86068-5}
+}
+
+@article{HuttRougier:2,
+author = {Hutt, Axel and P Rougier, Nicolas},
+year = {2013},
+month = {01},
+pages = {},
+title = {Numerical Simulation Scheme of One- and Two Dimensional Neural Fields Involving Space-Dependent Delays},
+journal = {Neural Fields: Theory and Applications},
+doi = {10.1007/978-3-642-54593-1_6}
+}
+
+@article{Simulator,
+author = {Nichols, Eric and Hutt, Axel},
+year = {2015},
+month = {09},
+pages = {25},
+title = {Neural Field Simulator: two-dimensional spatio-temporal dynamics involving finite transmission speed},
+volume = {9},
+journal = {Frontiers in Neuroinformatics},
+doi = {10.3389/fninf.2015.00025}
+}
+
+@article{Amari,
+author = {Amari, Shun-ichi},
+year = {1977},
+month = {09},
+pages = {77-87},
+title = {Dynamic of pattern formation in lateral-inhibition type neural fields},
+volume = {27},
+journal = {Biological cybernetics},
+doi = {10.1007/BF00337259}
+}
+
+@article{SequeiraLima,
+author={Sequeira, T and Lima, P},
+title={Numerical Simulations of One- and Two-dimensional Stochastic Neural Field Equations with Delay - Submitted},
+JOURNAL={Journal of Computational Neuroscience},
+YEAR={2021}
+}
+
+@article{Julia,
+author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral},
+year = {2015},
+month = {07},
+pages = {},
+title = {Julia: A Fresh Approach to Numerical Computing},
+doi = {10.1137/141000671}
+}
+
+@INPROCEEDINGS{Kulikov1D,
+author={Kulikov, Gennady Yu. and Kulikova, Maria V. and Lima, Pedro M.},
+booktitle={2019 23rd International Conference on System Theory, Control and Computing (ICSTCC)},
+title={Numerical Simulation of Neural Fields with Finite Transmission Speed and Random Disturbance},
+year={2019},
+pages={644-649},
+doi={10.1109/ICSTCC.2019.8885972}}
+
+@mastersthesis{Sequeira,
+author = {Sequeira, Tiago},
+title = {Numerical simulations of one- and two-dimensional stochastic neural field equations with delay},
+school = {Instituto Superior Técnico - Universidade de Lisboa},
+year = {2021},
+note = {\url{https://fenix.tecnico.ulisboa.pt/cursos/mma/dissertacao/1691203502344856}}}
+
+@article{Coombes:WavesBumps,
+author = {Coombes, Stephen},
+year = {2005},
+month = {09},
+pages = {91-108},
+title = {Waves, bumps, and patterns in neural field theories. Biol. Cybern. 93(2), 91-108},
+volume = {93},
+journal = {Biological cybernetics},
+doi = {10.1007/s00422-005-0574-y}
+}
+
[…]
