### 2022-03-01 #3974 @tiagoseq
Hello,
Yes, the new version of the package has been released. To update, please, go to julia terminal and type `] update`. One can see the new documentation and the status bar, while executing the function `solveNFE`.

However, the package is still too big. Even after moving the documentation to an independent repository, and after the creation of a new orphan branch and deleting the old master branch and commit history.

I have to check this question carefully. The problem relies on `.git` folder. I've tried another approaches but without success. Probably tomorrow I'll give some updates.

Nevertheless, the new version of the package can be checked after updating the local julia registries. If you have any issue updating the version 0.1.3 tell me.

Thank you!
