### 2022-03-02 #3974 @rougier
Thanks for the update. For the size of the repository, I think it's ok. The .git folder probably carry the whole history and this might explain the size. For future code, make sure to not add all the files blindly or the problem will reappear. I've checked the new PDF and it now looks good to me. @jarvist @gdetor Looks good to me.
