### 2022-03-02 #3974 @tiagoseq
Yes, learned my lesson. However, I still want to overcome this issue, to future downloads not to be so heavy.
Thank you for your advice to add a progress bar @rougier.
If i have any update about the commit history I'll let you know.
