### 2022-03-08 #3974 @tiagoseq
Hi @gdetor thank you for your help in improvig the manuscript and the source code!

In order to move on with this process, I think you must complete your check list, as @rougier did, at the top of this thread.

Thanks,
Tiago
