### 2022-03-10 #3974 @tiagoseq
I searched for an answer in [JOSS readthedocs](https://joss.readthedocs.io/en/latest/index.html) and didn't see anything with respect on how complete the checklist.

@jarvist can you give any help? Or maybe @rougier could give some insight on how complete the check list
