### 2022-03-23 #3974 @tiagoseq
> @jarvist @tiagoseq It seems I cannot access the check list. Any help?

Hi @gdetor , just to check, when you say you cannot access the check list in the beginning of this thread, you're saying that you could not edit the comment that contains the checklist, right?
It seems that @rougier didn't have any problem with this:
![image](https://user-images.githubusercontent.com/73851660/159727697-660e8c19-1fed-43e7-ba76-fcc2f8438ae8.png)

I can't edit the edit the check list either, because probably only the reviewers can.
When reading the initial comments, I found the following post:

> The reviewer already has a pending invite.
>
> @gdetor please accept the invite by clicking this link: https://github.com/openjournals/joss-reviews/invitations

Did you accepted this invitation? If not, probably this is the problem. If you accepted the invitation to be a reviewer then I guess only @jarvist can help us.

Thank you
