### 2022-03-24 #3974 @danielskatz
bot commands need to be the first thing in a comment, but also, `@whedon` has been replaced by `@editorialbot`, and our method for generating and using checklists has changed.  In this case, and for new reviews going forward, the reviewer (@gdetor) should put `@editorialbot generate my checklist` at the start of a new comment to get a checklist to use. This is a bit confusing here, since this review started under the old system.
