### 2022-03-25 #3974 @tiagoseq
Hi @gdetor, did you succeed in completing the checklist?

@danielskatz Thank you for your help! So, if the review started under the old system, what does it means in practice? That the other reviewer, @rougier, has to generate a new checklist with the @editorialbot ? Or it is ok to proceed the process with one reviewer with the "old" checklist and the other one with the checklist generated with the @editorialbot ?
