### 2022-03-25 #3974 @danielskatz
This can stay as is, with mixed checklists.

@gdetor might want to try the command again, as GitHub was dropping some requests for a bit yesterday.
