### 2022-04-09 #3974 @jarvist
@tiagoseq - I must admit I'm more than a little bit concerned at the fact that you appear to have deleted all git history & then force-pushed to your GitHub repository, since the start of this review. Was this intentional? Could you explain why?

I was just trying to check that it made sense for you to be the sole author on the paper, c.f. the contribution to the package, and the fact that you have your (academic?) supervisor in the acknowledgments only.
