### 2022-04-09 #3974 @tiagoseq
Hello @jarvist , yes, it was intentional. I was trying to address the issue that @rougier pointed out, regarding the repository size,
>     1. Repository is way too big (36Mo) due mostly to the presence of gif animations in the imag and doc directories. Problem is that the doc directories containes the build website for no obvious reason. I would advise to cleanup the repository. But problem will remain because the build addition has been commited and the repository will stay big unless the git history is rewritten. Another solution would be to create a new repository and make sure to only add necessary stuff.

As Rougier said, the large size was due to the gifs and animations that are present in the documentation.
To resolve this I created a new repository for the sole purpose of hosting the package documentation, [NFEdocs.jl](https://tiagoseq.github.io/NFEdocs.jl/dev/), then I deleted the commit history in an attempt of reducing the size of the repo.

On the question of being the only author of the paper, I wrote the entire julia package from scratch, since this is integrated in the scope of my master thesis I did not receive any input from other students or professors. Moreover, the entire manuscript was written by me and my academic supervisor read the manuscript and pointed out some typos, before I submitted it.

I hope the questions have been clarified!

Thank you,
Tiago
