### 2022-04-10 #3974 @jarvist
> To resolve this I created a new repository for the sole purpose of hosting the package documentation, [NFEdocs.jl](https://tiagoseq.github.io/NFEdocs.jl/dev/), then I deleted the commit history in an attempt of reducing the size of the repo.

Ah, OK! I understand the motivation now. But this is quite sub optimal, as it stops a future user / maintainer being able to `git bisect` or `git blame` while debugging, and otherwise understanding the construction of the package.
Presuming you still have the original repository, rewriting the history to remove the large files & then force-pushing the resulting repo to github would be ideal.
This kind of git stuff is a little bit scary & difficult. I think I used `git-filter-repo` the last time I needed to do this: https://htmlpreview.github.io/?https://github.com/newren/git-filter-repo/blob/docs/html/git-filter-repo.html
