### 2022-06-20 #3974 @tiagoseq
Hello,
After a few weeks of a hard physical recovery I'm finally able to address this issue.
First of all, I want to thank you all, @jarvist, @rougier and @gdetor for the patience shown throughout this issue.

Regarding the problem that @jarvist pointed out, possible future contributors could not do `git blame` or `git bisect` because the history has been erased. An attempt to reduce the size of the repo, due to the number of GIFs used in the documentation.

However, the only major change that the current version of the repository has in comparison to the previous one (the one that I deleted the commit history) is the docs are hosted in an independent [repository](https://github.com/tiagoseq/NFEdocs.jl), the other changes are minor, such as the addition of a status bar and clean up the function documentation.
The solver itself is pretty much the same as it was in the first released version, v.0.1.0.

Basically, the current version of the solver is the baseline and I (and future contributors) will add new features, new numerical methods and schemes, etc on top of this version of the code.
Actually, I will add a road map to the readme section to point out what are the next steps.

Since these `git` things are a bit scary for me and the difference between v.01.0 and the current v0.1.3, is basically the docs hosted in a different repo and a status bar, I think that for future contributors there will be no problem to trace back possible bugs and errors. The very core computations of the solver weren't changed since its first release.

Once again, thank you so much for your patience.
I'm waiting for your response!
Tiago
