### 2022-06-28 #3974 @jarvist
OK! Given the mitigating circumstances, and having read the code (and seen that it well commented and has good test coverage), I think this is acceptable.

I've just made a few suggested edits to your paper, for stylistic reasons:
https://github.com/tiagoseq/NeuralFieldEq.jl/pull/32
