### 2022-06-29 #3974 @tiagoseq
I've already generated a DOI for this version (using Zenodo): [DOI for v.0.1.3](https://doi.org/10.5281/zenodo.6226695)
