+++ JOSS/paper.bib
@@ -51,12 +51,21 @@ journal = {Biological cybernetics},
 doi = {10.1007/BF00337259}
 }
 
-@article{SequeiraLima,
-author={Sequeira, T and Lima, P},
-title={Numerical Simulations of One- and Two-dimensional Stochastic Neural Field Equations with Delay - Submitted},
-JOURNAL={Journal of Computational Neuroscience},
-YEAR={2021}
+@Article{Sequeira2022,
+author={Sequeira, Tiago F. and Lima, Pedro M.},
+title={Numerical simulations of one- and two-dimensional stochastic neural field equations with delay},
+journal={Journal of Computational Neuroscience},
+year={2022},
+month={Aug},
+day={01},
+volume={50},
+number={3},
+pages={299-311},
+issn={1573-6873},
+doi={10.1007/s10827-022-00816-w}
 }
+
+
 
 @article{Julia,
 author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral},
