### 2022-06-29 #3974 @tiagoseq
Done, the missing DOI is already inserted (and the correspondent .bib entry updated). Do I need to generate the paper with the new entry?
