+++ JOSS/paper.bib
@@ -51,7 +51,7 @@ journal = {Biological cybernetics},
 doi = {10.1007/BF00337259}
 }
 
-@Article{Sequeira2022,
+@Article{SequeiraLima,
 author={Sequeira, Tiago F. and Lima, Pedro M.},
 title={Numerical simulations of one- and two-dimensional stochastic neural field equations with delay},
 journal={Journal of Computational Neuroscience},
@@ -64,8 +64,6 @@ pages={299-311},
 issn={1573-6873},
 doi={10.1007/s10827-022-00816-w}
 }
-
-
 
 @article{Julia,
 author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral},
