### 2022-07-03 #3974 @tiagoseq
Hi @jarvist! I do not know why the DOI is still invalid, I've already updated the respective entry. I've copied the bib entry from the journal's website. The paper was published in June of this year, maybe this is the 'problem'?
